Move unit tests execution to stestr
From time to time, unit tests execution fails with error: "TypeError:
'NoneType' object is not iterable". It seems that problem is in
unittest2 library, which is used by testtools for executing the tests.
Stestr lib is usinig unittest by default and the problem does not appear
for several hundred launches.

Authored-By: danpawlik <dpawlik@redhat.com>

Change-Id: I6646629128d9265b5fb7390dc131c55cf34a07ca
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>