Paths no longer contain double '/'
Paths created trough formatted string included double '/'.
While still parsable, and as far as I know posix valid,
it is nonetheless a bit confusing.

This change merely removes one of the '/',
resulting in path that is just as valid, just easier on the eye.

Tests were adjusted to reflect the change.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ic6bf6d0a02bead00c0e3c5a858476cf5f7c0d2a7