New optional param for log path
As we now have a mechanism for falling back to
a default directory for logs and artifacts, we can allow
the customers to specify their own.

This patch add the argument to the CLI along with
minimum necessary information about the usage.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Id31a7ce2dfc4d9a132fea7d600e8beb8adc0739b
(cherry picked from commit 7876cab41586157a1b5e8fd426deb16d79567a52)