add spinner to highlight progress
Implementing a Spinner class and utilizing it around a run action.  This
will add a moving icon that animates in place until output is reached,
letting the operator know that things are happening.

The interactive TTY test is included to ensure that during some unknown
usage such as logging to disk this feature doesn't spew unnecessary
clutter.

Change-Id: Ieef256a5c12b238008e9250c0ee182d80a2b6dfb