Test cross contamination prevention
The existing tests were using shared fakes as return value
of certain mocks. Unfortunately, while perfectly fine in most cases,
certain methods change values passed in place.
This created a race condition, which caused fluctuation in the reported
coverage of the relevant modules.

The undesirable test behavior only becomes apparent when the tests
are run in a succession and their behavior is analyzed, for example
during the run of the recently implemented 'coverchange' job.

Final statistics of the coverage measurement, displaying the totals
for number of lines, lines without coverage, branches etc., display
following pattern. Note the third column.

TOTAL                        995    105    340     56    86%
TOTAL                        995    105    340     56    86%
TOTAL                        995    105    340     56    86%
TOTAL                        995    108    340     56    86%
TOTAL                        995    108    340     56    86%
TOTAL                        995    105    340     56    86%
TOTAL                        995    105    340     56    86%
TOTAL                        995    108    340     56    86%
TOTAL                        995    105    340     56    86%

Number of lines without coverage oscillates between '108' and '105' depending
on the order in which the tests were run. The culprit being one of the functions
of the 'validations_libs.cli.common' module, which edits data passed in place.

As the data processed are passed from patched object, and the data itself are
defined withing the relevant 'fakes' module, it emerges with inevitability,
that the tested code operates on data previously altered by the tests.

Furthermore the 'test_run_command_failed_validation' test was augmented
to place further assertion on the arguments passed.

Closes-Bug: #1935003

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I2210fa1775691dd503eb882a24824eaf18d7bd3e