Enforce existing roles and playbooks checks
This patch enforces the roles and playbooks checks when creating a new
community validation by checking first if the ANSIBLE_ROLES_DIR and
ANSIBLE_VALIDATION_DIR are existing in the host.

It avoids getting error message like:
  - [Errno 2] No such file or directory: '/usr/share/ansible/roles'
when validations-common and tripleo-validations are not installed.

It also adjusts the unittests in order to avoid leaving leftovers such
as community-validations directory during the test execution.

Change-Id: I6facca59c2ca9db2b5b1e7905ad8c9be68fff0b4
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>