Using UBI9 as base for VF development containers
Centos container file[0] has not been updated in considerable time
and is increasingly out of sync with environment VF is meant to interact with.
Moving to UBI9[1] should prevent potential issues of stemming
from aformentioned divergence.

Furthermore, several minor stylistic issues of the existing container file
were resolved, such as comments and uperfluous commands.
Consequently the version was incremented to 1.0.

[0]https://hub.docker.com/_/centos
[1]https://hub.docker.com/r/redhat/ubi9

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ied821c2a7210d8131e154020272c46048372af29