Removing Popen object return from utils.run_command_and_log
Argument 'retcode_only' of the function and associated logic was removed
as it was not beneficial to the purpose, or use of the function.

Before this patch, only call of the function setting 'retcode_only'
to 'True' was made in tests of the validations-libs. All other calls
worked with assumption that Bool, not object will be returned,
and didn't supply the argument explicitly. Instead relying on
the default value.

In addition, it is not clear how would the Popen[0] object be
used, if it was returned.

The function has limited impact across Openstack[1],
only calls to it are made within validations-libs, although there is
a homonymous function withing tripleoclient intended
for the same task.

Docstring and tests were adjusted accordingly.

[0]https://docs.python.org/3/library/subprocess.html#popen-objects
[1]https://codesearch.openstack.org/?q=run_command_and_log&i=nope&literal=nope&files=&excludeFiles=&repos=

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I3f9c2ef364de645a99d28c45c96560b78862c848