Restoring logging facility and implementing default logger
The logging statements in our modules, including CLI, are now
using a dedicated VF logger with Syslog and stdout handlers.

This will ensure, that all messages will be properly journaled
and that only messages >warn will be displayed to user by default.
Patch only changes behavior of the logging statements in the VF
code itself. Log statements in imported modules, such as ansible-runner,
will behave as before.

In the absence of available '/dev/log' device the user will be notified,
during logger init, by a warning level message.

Tests for the new code are included.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: If8ed6510dad16dc8495717789bb132b957828e0d