Check the presence of the output callback and the roles path
This patch check the presence of the output callback *vf_validation_json*
which is mandatory for a correct behavior of the VF.
Without this checks the VF will execute the ansible command,
but wont be able to find the json log associated to the run,
and will failed at the end.
This check prevent useless investigations and save time for the user.

It also check the presence of the roles path (if one of all the paths is correct)
to prevent as well strange behavior from Ansible and the framework.

Change-Id: Iea6fc17ec4da91df652be0a0cb9338e3c512f141